company service add update and delete

+++ crm-app/src/main/java/com/hotelcrm/crmapp/repository/CompanyRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CompanyRepository extends JpaRepository<Company, Integer>, JpaSpecificationExecutor<Company> {
+public interface CompanyRepository extends JpaRepository<Company, Long>, JpaSpecificationExecutor<Company> {
 
     @Query("""
             SELECT new com.hotelcrm.crmapp.dto.CompanySummaryDto(
